Vtable fix.

+++ include/common/GDocView.h
@@ -373,7 +373,7 @@ public:
 	virtual ssize_t IndexAt(int x, int y) { return 0; }
 
 	/// Index=-1 returns the x,y of the cursor, Index >=0 returns the specified x,y
-	virtual bool GetLineColumnAtIndex(GdcPt2 &Pt, int Index = -1) { return false; }
+	virtual bool GetLineColumnAtIndex(GdcPt2 &Pt, ssize_t Index = -1) { return false; }
 
 	/// True if the document has changed
 	virtual bool IsDirty() { return false; }
